Fix the typo in class name Symfony, not Symphony
